Standardise drawing of SVGs
