+++ BnbnavNetClient/Views/MapView.axaml.cs
@@ -9,9 +9,7 @@ using System;
 using System.Linq;
 using System.Reactive;
 using BnbnavNetClient.Services;
-using Avalonia.Platform;
 using Avalonia.Svg.Skia;
-using Svg.Skia;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Diagnostics.CodeAnalysis;
@@ -38,13 +36,10 @@ public partial class MapView : UserControl
     Matrix _toScreenMtx = Matrix.Identity;
     Matrix _toWorldMtx = Matrix.Identity;
 
-    readonly IAssetLoader _assetLoader;
-
     public MapViewModel MapViewModel => (MapViewModel)DataContext!;
 
     public MapView()
     {
-        _assetLoader = AvaloniaLocator.Current.GetService<IAssetLoader>()!;
         _i18n = AvaloniaLocator.Current.GetRequiredService<IAvaloniaI18Next>();
 
         InitializeComponent();
@@ -283,7 +278,6 @@ public partial class MapView : UserControl
         }));
     }
 
-    readonly Dictionary<string, SKSvg> _svgCache = new();
     readonly IAvaloniaI18Next _i18n;
 
     List<(Point, Point, Edge)> DrawnEdges { get; set; } = new();
@@ -422,43 +416,8 @@ public partial class MapView : UserControl
         {
             return;
         }
-
-        SKSvg? svg = null;
-
-        var uriString = $"avares://BnbnavNetClient/Assets/Landmarks/{landmark.Type}.svg";
-        if (_svgCache.TryGetValue(uriString, out var outSvg))
-        {
-            svg = outSvg;
-        }
-        else
-        {
-            var uri = new Uri(uriString);
-            if (_assetLoader.Exists(uri))
-            {
-                var asset = _assetLoader.Open(uri);
-
-                svg = new();
-                svg.Load(asset);
-                if (svg.Picture is null)
-                    return;
-                _svgCache.Add(uriString, svg);
-            }
-        }
-
-        if (svg is null)
-            return;
-
-        var sourceSize = new Size(svg.Picture!.CullRect.Width, svg.Picture.CullRect.Height);
-        var scaleMatrix = Matrix.CreateScale(
-            rect.Width / sourceSize.Width,
-            rect.Height / sourceSize.Height);
-        var translateMatrix = Matrix.CreateTranslation(
-            rect.X * sourceSize.Width / rect.Width,
-            rect.Y * sourceSize.Height / rect.Height);
-
-        using (context.PushClip(rect))
-        using (context.PushPreTransform(translateMatrix * scaleMatrix))
-            context.Custom(new SvgCustomDrawOperation(rect, svg));
+        
+        context.DrawSvgUrl($"avares://BnbnavNetClient/Assets/Landmarks/{landmark.Type}.svg", rect);
     }
 
     [SuppressMessage("ReSharper", "PossibleLossOfFraction")]
@@ -506,55 +465,12 @@ public partial class MapView : UserControl
 
         foreach (var player in MapViewModel.MapService.Players.Values)
         {
-            SKSvg? svg = null;
-
-            var uriString = $"avares://BnbnavNetClient/Assets/playermark.svg";
-            if (_svgCache.TryGetValue(uriString, out var outSvg))
-            {
-                svg = outSvg;
-            }
-            else
-            {
-                var uri = new Uri(uriString);
-                if (_assetLoader.Exists(uri))
-                {
-                    var asset = _assetLoader.Open(uri);
-
-                    svg = new();
-                    svg.Load(asset);
-                    if (svg.Picture is null)
-                        return;
-                    _svgCache.Add(uriString, svg);
-                }
-            }
-
-            if (svg is null)
-                return;
 
             const int playerSize = 48;
             var rect = new Rect(ToScreen(player.MarkerCoordinates) - new Point(playerSize, playerSize) / 2, new Size(playerSize, playerSize));
+            const string uriString = "avares://BnbnavNetClient/Assets/playermark.svg";
+            context.DrawSvgUrl(uriString, rect, -player.MarkerAngle);
 
-            var sourceSize = new Size(svg.Picture!.CullRect.Width, svg.Picture.CullRect.Height);
-            var scaleMatrix = Matrix.CreateScale(
-                rect.Width / sourceSize.Width,
-                rect.Height / sourceSize.Height);
-            var translateMatrix = Matrix.CreateTranslation(
-                rect.X * sourceSize.Width / rect.Width,
-                rect.Y * sourceSize.Height / rect.Height);
-            var rotateMatrix = Matrix.CreateRotation(-player.MarkerAngle * double.Tau / 360.0);
-            var preRotateMatrix = Matrix.CreateTranslation(-sourceSize.Width / 2, -sourceSize.Width / 2);
-
-            // context.DrawEllipse(new SolidColorBrush(player.SnappedEdge is null ? new Color(255, 255, 0, 0) : new Color(255, 0, 255, 0)), null, ToScreen(player.MarkerCoordinates), 50, 50);
-            // context.DrawLine(new Pen(new SolidColorBrush(new Color(255, 0, 255, 0))), ToScreen(player.Velocity.Point1), ToScreen(player.Velocity.Point2));
-            
-            using (context.PushClip(rect))
-            using (context.PushPreTransform(translateMatrix * scaleMatrix))
-            using (context.PushTransformContainer())
-            using (context.PushPostTransform(preRotateMatrix))
-            using (context.PushPostTransform(rotateMatrix))
-            using (context.PushPostTransform(preRotateMatrix.Invert()))
-                context.Custom(new SvgCustomDrawOperation(rect, svg));
-            
             //Draw the player name
             var textBrush = (Brush)this.FindResource("ForegroundBrush")!;
